simd refactor
