+++ lib/segment/src/lib.rs
@@ -1,3 +1,5 @@
+#![feature(stdsimd)]
+
 mod common;
 pub mod entry;
 pub mod fixtures;
